@@ -1,46 +1,44 @@
 {
-  "name": "nergal.aws.com",
-  "version": "0.1.2",
-  "private": true,
-  "dependencies": {
-    "@testing-library/jest-dom": "^5.11.6",
-    "@testing-library/react": "^11.2.2",
-    "@testing-library/user-event": "^12.5.0",
-    "@types/jest": "^26.0.19",
-    "@types/lodash": "^4.14.165",
-    "@types/node": "^14.14.12",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "aws-amplify": "^3.3.12",
-    "bootstrap": "^4.5.3",
-    "lodash": "^4.17.20",
-    "node-sass": "^4.14.1",
-    "react": "^17.0.1",
-    "react-bootstrap": "^1.4.0",
-    "react-dom": "^17.0.1",
-    "react-ga": "^3.3.0",
-    "react-scripts": "^4.0.1",
-    "react-string-replace": "0.4.4",
-    "typescript": "^3.9.7"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test"
-  },
-  "eslintConfig": {
-    "extends": "react-app"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  }
+	"name": "nergal.cv",
+	"version": "0.2.0",
+	"private": true,
+	"dependencies": {
+		"@testing-library/jest-dom": "^6.6.3",
+		"@testing-library/react": "^16.0.1",
+		"@testing-library/user-event": "^14.5.2",
+		"@types/jest": "^29.5.14",
+		"@types/lodash": "^4.17.13",
+		"@types/node": "^22.10.1",
+		"@types/react": "^18.3.12",
+		"@types/react-dom": "^18.3.1",
+		"jotai": "^2.10.3",
+		"lodash": "^4.17.21",
+		"next": "latest",
+		"react": "^18.3.1",
+		"react-dom": "^18.3.1",
+		"react-ga": "^3.3.1",
+		"react-string-replace": "1.1.1",
+		"typescript": "^5.7.2"
+	},
+	"scripts": {
+		"dev": "next",
+		"build": "next build",
+		"start": "next start",
+		"lint": "npx --yes @biomejs/biome ci"
+	},
+	"browserslist": {
+		"production": [">0.2%", "not dead", "not op_mini all"],
+		"development": [
+			"last 1 chrome version",
+			"last 1 firefox version",
+			"last 1 safari version"
+		]
+	},
+	"devDependencies": {
+		"@biomejs/biome": "1.9.4",
+		"autoprefixer": "^10.4.20",
+		"postcss": "^8.4.49",
+		"sass": "^1.81.0",
+		"tailwindcss": "^3.4.15"
+	}
 }
