@@ -1,7 +1,7 @@
-import General from "./General";
-import Sidebar from "./Sidebar";
-import Overview from "./Overview";
-import Additionally from "./Additionally";
-import Experience from "./Experience";
+import { Additionally } from "./Additionally";
+import { Experience } from "./Experience";
+import { General } from "./General";
+import { Overview } from "./Overview";
+import { Sidebar } from "./Sidebar";
 
 export { General, Sidebar, Overview, Additionally, Experience };
