@@ -1,35 +1,29 @@
-import React, { FC, useCallback } from "react";
-import { Analytics } from "aws-amplify";
+import React, { type FC } from "react";
 
 type Props = {
-  name: string;
-  url?: string;
+	name: string;
+	url?: string;
 };
 
-const VendorBadge: FC<Props> = ({ name, url }) => {
-  const onLinkClick = useCallback(() => {
-    Analytics.record({ name: "vendor.link", value: url });
-  }, [url]);
+export const VendorBadge: FC<Props> = ({ name, url }) => {
+	if (!name) return null;
 
-  if (!name) return null;
+	if (url) {
+		return (
+			<span className="normal-case">
+				(at{" "}
+				<a
+					className="underline decoration-1 underline-offset-4 decoration-black decoration-dotted text-blue-500 hover:text-blue-800"
+					target="_blank"
+					rel="noopener noreferrer"
+					href={url}
+				>
+					{name}
+				</a>
+				)
+			</span>
+		);
+	}
 
-  if (url) {
-    return (
-      <span className="vendor">
-        (at{" "}
-        <a
-          target="_blank"
-          rel="noopener noreferrer"
-          onClick={onLinkClick}
-          href={url}
-        >
-          {name}
-        </a>
-        )
-      </span>
-    );
-  }
-  return <span className="vendor">(at {name})</span>;
+	return <span className="normal-case">(at {name})</span>;
 };
-
-export default VendorBadge;
